chore: move options to its own file
